% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_proxy.R
\name{Proxy}
\alias{Proxy}
\title{Drug Proxy Class}
\usage{
Proxy(
  drug_name,
  gene_name,
  gene_chr,
  gene_start,
  gene_end,
  gene_flanks_kb,
  clump_p1,
  clump_p2,
  clump_r2,
  clump_kb
)
}
\arguments{
\item{drug_name}{.}

\item{gene_name}{.}

\item{gene_chr}{.}

\item{gene_start}{.}

\item{gene_end}{.}

\item{gene_flanks_kb}{.}

\item{clump_p1}{.}

\item{clump_p2}{.}

\item{clump_r2}{.}

\item{clump_kb}{.}
}
\value{
a Proxy S3 object
}
\description{
An S3 class holding default settings for a drug proxy instrument.
}
